// src/backend/models/profile.ts
export class Profile {
  fullname?: string;
  birthdate?: string;
  familyName?: string;
  gender?: string;
  givenName?: string;
  locale?: string;
  middleName?: string;
  name?: string;
  nickname?: string;
  picture?: string;
  preferredUsername?: string;
  website?: string;
  address?: string;
  avatar?: string;
  description?: string;
  job?: string;
}